requirements.txt
dagster==1.6.7
dagster-mongo==0.20.7
pymongo==4.6.1
pandas==2.2.1
python-dotenv==1.0.1
pytest==8.0.2 